Use submittedAt, pin nextSteps enum, add endpoint tests
